Some minor improvements
Changed the name of Comparison.apply to convert. This is because, if the underlying type T is Int, the compiler gets confused with apply(x) when it was intended it be part of apply(x)(y). This is essentially an internal change and doesn't really affect the API.
Comparer: cosmetic change to remove braces from lambdas.